+++ lib/squadron_annotations.dart
@@ -1,2 +1,3 @@
+export 'src/annotations/squadron_serializer.dart';
 export 'src/annotations/squadron_service.dart';
 export 'src/annotations/squadron_method.dart';